app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "InvisID - Leak Attribution System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.2",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
    "numpy>=1.26.0",
    "pycryptodome>=3.18.0",
    "invisible-watermark>=0.2.0",
    "torch",
    "torchvision",
    "fpdf2>=2.8.7",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "httpx>=0.28.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Enable security (S), bugbear (B), and complexity (C90) rules
select = ["E", "F", "B", "S", "C90", "I"]
ignore = ["B008"] # Allow FastAPI Depends() and File() calls in arguments

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"] # Allow asserts in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
